+++ mediakit-lite/functions.php
@@ -13,7 +13,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 /**
  * Define Constants
  */
-define( 'MKP_THEME_VERSION', '1.2.2' );
+define( 'MKP_THEME_VERSION', '1.2.3' );
 define( 'MKP_THEME_DIR', get_template_directory() );
 define( 'MKP_THEME_URI', get_template_directory_uri() );
 
@@ -208,7 +208,9 @@ if ( defined( 'WP_DEBUG' ) && WP_DEBUG ) {
 
 // Ensure all files exist before requiring them
 $required_files = array(
+    '/inc/auto-contrast.php',
     '/inc/customizer.php',
+    '/inc/customizer-dynamic-styles.php',
     '/inc/template-tags.php',
     '/inc/template-functions.php',
     '/inc/field-helpers.php',